fix: tests for TS
