+++ package.json
@@ -38,9 +38,11 @@
   "devDependencies": {
     "@tsconfig/node14": "^1.0.1",
     "@types/node": "^17.0.33",
+    "@types/tap": "^15.0.7",
     "fastify": "^3.15.1",
     "snazzy": "^9.0.0",
     "tap": "^16.2.0",
+    "ts-node": "^10.7.0",
     "ts-standard": "^11.0.0",
     "typescript": "~4.4.4"
   },
